Thumbtack-Internship/TalkToYourSanta/frontend/src/app/components/header/header.component.ts
import {Component, OnInit} from '@angular/core';
import {animate, style, transition, trigger} from '@angular/animations';

@Component({
  selector: 'app-header',
  templateUrl: './header.component.html',
  styleUrls: ['./header.component.sass'],
  animations: [
    trigger(
      'inOutAnimation',
      [
        transition(
          ':enter',
          [
            style({ opacity: 0 }),
            animate('0.2s', style({opacity: 1 }))
          ]
        ),
        transition(
          ':leave',
          [
            style({ opacity: 1 }),
            animate('0.2s', style({ opacity: 0 }))
          ]
        )
      ]
    )
  ]
})
export class HeaderComponent implements OnInit {

  chatName: string;
  showMenu = false;

  constructor() { }

  ngOnInit(): void {
    this.chatName = 'Santa';
  }

  toggleMenu(): void {
    this.showMenu = !this.showMenu;
  }
}
